touching the Makefile to reset GitHub's CDN
It looks like Github optimized their README's that contain images to use
a CDN. Since I made a mistake the first time with my image, I think it
froze or something, since the URL seems to contain a hash, but it
doesn't point to the image, which is now properly hosted. I'm going to
just try committing, then uncommitting, and see if it changes things.

Bad form if this were a real public repo, I know, but this is hardly a
monumental repository :-p